+++ src/test/resources/pgdiff_test_files/alter_schema_diff.sql
@@ -1,0 +1,1 @@
+ALTER SCHEMA blah OWNER TO asi;